Query functions now return tag functions
Mostly this means functions like ``db.query(SQL`SELECT *`);``
can now be ``db.query()`SELECT *`;``

This is a bit less noisy. We still require `()` since most tag
functions do not have side effects but these definitely do.

+++ src/test/actions.test.ts
@@ -109,9 +109,8 @@ describe('Loginserver actions', () => {
 	describe('Ladder', () => {
 		it('Should update the ladder', async () => {
 			for (const id of ['catra', 'adora']) {
-				await tables.ladder.deleteOne(
-					SQL`userid = ${id} AND formatid = ${'gen1randombattle'}`,
-				); // clear their ratings entirely
+				// clear their ratings entirely
+				await tables.ladder.deleteOne()`userid = ${id} AND formatid = ${'gen1randombattle'}`;
 			}
 			const {result} = await utils.testDispatcher({
 				act: 'ladderupdate',
@@ -131,9 +130,7 @@ describe('Loginserver actions', () => {
 			const p1 = NTBBLadder.getUserData('shera')!;
 			const p2 = NTBBLadder.getUserData('catra')!;
 			for (const player of [p1, p2]) {
-				await tables.ladder.deleteAll(
-					SQL`WHERE userid = ${player.id} AND formatid = ${ladder.formatid}`,
-				);
+				await tables.ladder.deleteAll()`WHERE userid = ${player.id} AND formatid = ${ladder.formatid}`;
 			}
 			await ladder.updateRating(p1, p2, 1);
 			const {result} = await utils.testDispatcher({
